Move handleObjData/handleStlData/handleXyzData into formatDataHandlers.ts
Joins the other per-format data handlers already extracted there since
they follow the same shape (parse message.data, build SpatialData,
addNewFiles/displayFiles).
