+++ engine/src/formatDataHandlers.ts
@@ -1,5 +1,5 @@
 import * as THREE from 'three';
-import { SpatialData } from './interfaces';
+import { SpatialData, SpatialFace, SpatialVertex } from './interfaces';
 
 export interface FormatDataHandlersHost {
   scene: THREE.Scene;
@@ -394,6 +394,347 @@ export async function handleXyzVariantData(
     console.error('Error handling XYZ variant data:', error);
     host.showError(
       `${message.variant.toUpperCase()} processing failed: ${error instanceof Error ? error.message : String(error)}`
+    );
+  }
+}
+
+export async function handleObjData(host: FormatDataHandlersHost, message: any): Promise<void> {
+  try {
+    console.log(`Load: recv OBJ ${message.fileName}`);
+    host.showStatus(`OBJ: processing ${message.fileName}`);
+
+    const objData = message.data;
+    const hasFaces = objData.faceCount > 0;
+    const hasLines = objData.lineCount > 0;
+    const hasPoints = objData.pointCount > 0;
+
+    console.log(
+      `OBJ: v=${objData.vertexCount}, pts=${objData.pointCount}, f=${objData.faceCount}, lines=${objData.lineCount}, groups=${objData.materialGroups ? objData.materialGroups.length : 0}`
+    );
+
+    // Convert OBJ vertices to PLY format
+    const vertices: SpatialVertex[] = objData.vertices.map((v: any) => ({
+      x: v.x,
+      y: v.y,
+      z: v.z,
+      red: 128, // Default gray color
+      green: 128,
+      blue: 128,
+    }));
+
+    // Convert OBJ faces to PLY format if they exist
+    const faces: SpatialFace[] = [];
+    if (hasFaces) {
+      for (const objFace of objData.faces) {
+        if (objFace.indices.length >= 3) {
+          faces.push({
+            indices: objFace.indices,
+          });
+        }
+      }
+    }
+
+    // Create PLY data structure
+    const spatialData: SpatialData = {
+      vertices,
+      faces,
+      format: 'ascii',
+      version: '1.0',
+      comments: [`Converted from OBJ file: ${message.fileName}`],
+      vertexCount: vertices.length,
+      faceCount: faces.length,
+      hasColors: true,
+      hasNormals: objData.hasNormals,
+      fileName: message.fileName, // Keep original OBJ filename
+      shortPath: message.shortPath,
+      fileIndex: host.spatialFiles.length,
+      fileSizeInBytes: message.fileSizeInBytes,
+    };
+
+    // Store OBJ-specific data for enhanced rendering
+    (spatialData as any).objData = objData;
+    (spatialData as any).isObjFile = true;
+    (spatialData as any).objRenderType = hasFaces ? 'mesh' : 'wireframe';
+
+    // Store line data for wireframe rendering (either as primary or secondary visualization)
+    if (hasLines) {
+      (spatialData as any).objLines = objData.lines;
+      (spatialData as any).hasWireframe = true;
+    }
+
+    // Store point data for point rendering
+    if (hasPoints) {
+      (spatialData as any).objPoints = objData.points;
+      (spatialData as any).hasPoints = true;
+    }
+
+    // Add to visualization
+    if (message.isAddFile) {
+      host.addNewFiles([spatialData]);
+    } else {
+      await host.displayFiles([spatialData]);
+    }
+
+    // Status message based on what was loaded
+    const statusParts = [`${vertices.length.toLocaleString()} vertices`];
+    if (hasPoints) {
+      statusParts.push(`${objData.pointCount} points`);
+    }
+    if (hasFaces) {
+      statusParts.push(`${faces.length.toLocaleString()} faces`);
+    }
+    if (hasLines) {
+      statusParts.push(`${objData.lineCount} line segments`);
+    }
+    if (objData.hasTextures) {
+      statusParts.push(`${objData.textureCoordCount} texture coords`);
+    }
+    if (objData.hasNormals) {
+      statusParts.push(`${objData.normalCount} normals`);
+    }
+
+    host.showStatus(`OBJ ${hasFaces ? 'mesh' : 'wireframe'} loaded: ${statusParts.join(', ')}`);
+  } catch (error) {
+    console.error('Error handling OBJ data:', error);
+    host.showError(
+      `Failed to process OBJ file: ${error instanceof Error ? error.message : String(error)}`
+    );
+  }
+}
+
+export async function handleStlData(host: FormatDataHandlersHost, message: any): Promise<void> {
+  try {
+    console.log(`Load: recv STL ${message.fileName}`);
+    host.showStatus(`STL: processing ${message.fileName}`);
+
+    const stlData = message.data;
+    const hasColors = stlData.hasColors;
+
+    console.log(
+      `STL: ${stlData.triangleCount} triangles, format=${stlData.format}, colors=${hasColors}`
+    );
+
+    // Handle empty STL files
+    if (stlData.triangleCount === 0 || !stlData.triangles || stlData.triangles.length === 0) {
+      console.log('STL: Empty mesh detected');
+      host.showStatus(`STL: Empty mesh loaded (${message.fileName})`);
+
+      // Create minimal PLY data for empty mesh
+      const spatialData: SpatialData = {
+        vertices: [],
+        faces: [],
+        format: stlData.format === 'binary' ? 'binary_little_endian' : 'ascii',
+        version: '1.0',
+        comments: [
+          `Empty STL mesh: ${message.fileName}`,
+          `Original format: ${stlData.format}`,
+          ...(stlData.header ? [`Header: ${stlData.header}`] : []),
+        ],
+        vertexCount: 0,
+        faceCount: 0,
+        hasColors: false,
+        hasNormals: false,
+        fileName: message.fileName.replace(/\.stl$/i, '_empty.ply'),
+        shortPath: message.shortPath,
+        fileIndex: host.spatialFiles.length,
+        fileSizeInBytes: message.fileSizeInBytes,
+      };
+
+      // Add to visualization (even empty files should be tracked)
+      if (message.isAddFile) {
+        host.addNewFiles([spatialData]);
+      } else {
+        await host.displayFiles([spatialData]);
+      }
+
+      return;
+    }
+
+    // Convert STL triangles to PLY vertices and faces
+    const vertices: SpatialVertex[] = [];
+    const faces: SpatialFace[] = [];
+    const vertexMap = new Map<string, number>(); // For deduplication
+
+    let vertexIndex = 0;
+
+    for (let i = 0; i < stlData.triangles.length; i++) {
+      const triangle = stlData.triangles[i];
+      const faceIndices: number[] = [];
+
+      // Process each vertex of the triangle
+      for (let j = 0; j < 3; j++) {
+        const vertex = triangle.vertices[j];
+        const key = `${vertex.x},${vertex.y},${vertex.z}`;
+
+        let vIndex = vertexMap.get(key);
+        if (vIndex === undefined) {
+          // New vertex
+          vIndex = vertexIndex++;
+          vertexMap.set(key, vIndex);
+
+          const plyVertex: SpatialVertex = {
+            x: vertex.x,
+            y: vertex.y,
+            z: vertex.z,
+            nx: triangle.normal.x,
+            ny: triangle.normal.y,
+            nz: triangle.normal.z,
+          };
+
+          // Add color if available
+          if (hasColors && triangle.color) {
+            plyVertex.red = triangle.color.red;
+            plyVertex.green = triangle.color.green;
+            plyVertex.blue = triangle.color.blue;
+          } else {
+            // Default gray color
+            plyVertex.red = 180;
+            plyVertex.green = 180;
+            plyVertex.blue = 180;
+          }
+
+          vertices.push(plyVertex);
+        }
+
+        faceIndices.push(vIndex);
+      }
+
+      // Add the face
+      faces.push({
+        indices: faceIndices,
+      });
+    }
+
+    // Create PLY data structure
+    const spatialData: SpatialData = {
+      vertices,
+      faces,
+      format: stlData.format === 'binary' ? 'binary_little_endian' : 'ascii',
+      version: '1.0',
+      comments: [
+        `Converted from STL file: ${message.fileName}`,
+        `Original format: ${stlData.format}`,
+        `Triangle count: ${stlData.triangleCount}`,
+        ...(stlData.header ? [`Header: ${stlData.header}`] : []),
+      ],
+      vertexCount: vertices.length,
+      faceCount: faces.length,
+      hasColors: true,
+      hasNormals: true,
+      fileName: message.fileName.replace(/\.stl$/i, '_mesh.ply'),
+      shortPath: message.shortPath,
+      fileIndex: host.spatialFiles.length,
+      fileSizeInBytes: message.fileSizeInBytes,
+    };
+
+    // Store STL-specific data for enhanced rendering
+    (spatialData as any).stlData = stlData;
+    (spatialData as any).isStlFile = true;
+    (spatialData as any).stlFormat = stlData.format;
+    (spatialData as any).stlTriangleCount = stlData.triangleCount;
+
+    // Add to visualization
+    if (message.isAddFile) {
+      host.addNewFiles([spatialData]);
+    } else {
+      await host.displayFiles([spatialData]);
+    }
+
+    // Status message
+    const statusParts = [
+      `${vertices.length.toLocaleString()} vertices`,
+      `${faces.length.toLocaleString()} triangles`,
+      `${stlData.format} format`,
+    ];
+    if (hasColors) {
+      statusParts.push('with colors');
+    }
+
+    host.showStatus(`STL mesh loaded: ${statusParts.join(', ')}`);
+  } catch (error) {
+    console.error('Error handling STL data:', error);
+    host.showError(
+      `Failed to process STL file: ${error instanceof Error ? error.message : String(error)}`
+    );
+  }
+}
+
+export async function handleXyzData(host: FormatDataHandlersHost, message: any): Promise<void> {
+  try {
+    console.log('Received XYZ data for processing:', message.fileName);
+    host.showStatus('Parsing XYZ file...');
+
+    // Parse XYZ file (simple format: x y z [r g b] per line)
+    const decoder = new TextDecoder('utf-8');
+    const text = decoder.decode(message.data);
+    const lines = text.split('\n').filter(line => line.trim().length > 0);
+
+    const vertices: SpatialVertex[] = [];
+    let hasColors = false;
+
+    for (const line of lines) {
+      const parts = line.trim().split(/\s+/);
+      if (parts.length >= 3) {
+        const x = parseFloat(parts[0]);
+        const y = parseFloat(parts[1]);
+        const z = parseFloat(parts[2]);
+
+        if (!isNaN(x) && !isNaN(y) && !isNaN(z)) {
+          const vertex: SpatialVertex = { x, y, z };
+
+          // Check for color data (RGB values)
+          if (parts.length >= 6) {
+            const r = parseInt(parts[3]);
+            const g = parseInt(parts[4]);
+            const b = parseInt(parts[5]);
+
+            if (!isNaN(r) && !isNaN(g) && !isNaN(b)) {
+              vertex.red = Math.max(0, Math.min(255, r));
+              vertex.green = Math.max(0, Math.min(255, g));
+              vertex.blue = Math.max(0, Math.min(255, b));
+              hasColors = true;
+            }
+          }
+
+          vertices.push(vertex);
+        }
+      }
+    }
+
+    if (vertices.length === 0) {
+      throw new Error('No valid vertices found in XYZ file');
+    }
+
+    // Create PLY data structure
+    const spatialData: SpatialData = {
+      vertices,
+      faces: [],
+      format: 'ascii',
+      version: '1.0',
+      comments: [`Converted from XYZ file: ${message.fileName}`],
+      vertexCount: vertices.length,
+      faceCount: 0,
+      hasColors,
+      hasNormals: false,
+      fileName: message.fileName.replace(/\.xyz$/i, '_pointcloud.ply'),
+      shortPath: message.shortPath,
+      fileIndex: host.spatialFiles.length,
+    };
+
+    // Add to visualization
+    if (message.isAddFile) {
+      host.addNewFiles([spatialData]);
+    } else {
+      await host.displayFiles([spatialData]);
+    }
+
+    host.showStatus(
+      `XYZ file loaded successfully! ${vertices.length.toLocaleString()} points${hasColors ? ' with colors' : ''}`
+    );
+  } catch (error) {
+    console.error('Error handling XYZ data:', error);
+    host.showError(
+      `Failed to process XYZ file: ${error instanceof Error ? error.message : String(error)}`
     );
   }
 }
